Implemented the NSColor Object within a Function for NSColorWell's Color Property
The reasons for this is because NSColorWell's property is of type NSColor in which NSColor is initialized by three CGFloat objects.  Once we programmed the logic for the initializer based on the NSSlider's doubleValue property, we created a method called updateColor() and inside that method we set the ColorWell color property by setting it to the initialized NSColor.